Implementación funcionaldiad admin y operador parcial

+++ frontend/src/app/components/admin/admin.component.ts
@@ -1,4 +1,3 @@
-// admin.component.ts
 import { Component, OnInit } from '@angular/core';
 import { UserService } from '../../services/user.service';
 import { AuthService } from '../../services/auth.service';
@@ -28,6 +27,7 @@ export class AdminComponent implements OnInit {
 
   loadUsers() {
     this.userService.getUsers().subscribe(data => {
+      console.log(data); // Verifica que los datos tengan un campo `id` o `_id`
       this.users = data;
     });
   }
@@ -39,13 +39,24 @@ export class AdminComponent implements OnInit {
   }
 
   deleteUser(id: string) {
+    if (!id) {
+      console.error('ID is undefined');
+      return;
+    }
+    console.log(id);
     this.userService.deleteUser(id).subscribe(() => {
       this.loadUsers();
     });
   }
 
   updateUser(user: any) {
+    console.log(user);
+    if (!user._id) {
+      console.error('User ID is undefined');
+      return;
+    }
     this.userService.updateUser(user).subscribe(() => {
+      console.log(user);
       this.loadUsers();
     });
   }
